WIP makind CRAM valid input/output
